+++ rust/avdschema/src/utils/load.rs
@@ -83,15 +83,10 @@ where
 #[cfg(test)]
 mod tests {
     use super::Load;
+    use crate::Store;
     use crate::any::AnySchema;
-    use crate::utils::dump::tests::{
-        dump_avd_store, dump_json, dump_store_yaml, dump_xz2, dump_yaml,
-    };
-    use crate::utils::test_utils::{
-        EOS_CLI_CONFIG_GEN_FRAGMENTS, EOS_DESIGNS_FRAGMENTS, get_test_dict_schema, get_test_store,
-        get_tmp_file,
-    };
-    use crate::{LoadFromFragments as _, Store, resolve_schema};
+    use crate::utils::dump::tests::{dump_json, dump_store_yaml, dump_xz2, dump_yaml};
+    use crate::utils::test_utils::{get_test_dict_schema, get_test_store, get_tmp_file};
 
     #[test]
     fn load_yaml() {
@@ -125,35 +120,6 @@ mod tests {
         dump_store_yaml();
         let file_path = get_tmp_file("test_dump_store.yml");
         let store = get_test_store();
-        let result = Store::from_file(Some(file_path));
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), store);
-    }
-    #[test]
-    pub(crate) fn load_avd_store() {
-        dump_avd_store();
-
-        // Load schemas from fragments, resolve all $ref and save in a store we can compare the loaded store with.
-        let mut eos_cli_config_gen_schema =
-            AnySchema::from_fragments(EOS_CLI_CONFIG_GEN_FRAGMENTS.into()).unwrap();
-        let mut eos_designs_schema =
-            AnySchema::from_fragments(EOS_DESIGNS_FRAGMENTS.into()).unwrap();
-        let mut store = Store {
-            eos_cli_config_gen: eos_cli_config_gen_schema.to_owned(),
-            eos_designs: eos_designs_schema.to_owned(),
-        };
-        resolve_schema(&mut eos_cli_config_gen_schema, &store).unwrap();
-        store.eos_cli_config_gen = eos_cli_config_gen_schema;
-        resolve_schema(&mut eos_designs_schema, &store).unwrap();
-        store.eos_designs = eos_designs_schema;
-
-        // Now load the previously dumped files and compare
-        let file_path = get_tmp_file("test_dump_avd_store_resolved.yml");
-        let result = Store::from_file(Some(file_path));
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), store);
-
-        let file_path = get_tmp_file("test_dump_avd_store_resolved.xz2");
         let result = Store::from_file(Some(file_path));
         assert!(result.is_ok());
         assert_eq!(result.unwrap(), store);
